update A[] datatype to be double precision floating point

+++ sum_direct.cpp
@@ -7,15 +7,14 @@
 #include <string.h>
 
 
-
 void 
-setup(int64_t N, uint64_t A[])
+setup(int64_t N, double A[])
 {
    printf(" inside direct_sum problem_setup, N=%lld \n", N);
 }
 
-int64_t
-sum(int64_t N, uint64_t A[])
+double
+sum(int64_t N, double A[])
 {
    printf(" inside direct_sum perform_sum, N=%lld \n", N);
