+++ sum_indirect.cpp
@@ -9,13 +9,13 @@
 
 
 void 
-setup(int64_t N, uint64_t A[])
+setup(int64_t N, double A[])
 {
    printf(" inside sum_indirect problem_setup, N=%lld \n", N);
 }
 
-int64_t
-sum(int64_t N, uint64_t A[])
+double
+sum(int64_t N, double A[])
 {
    printf(" inside sum_indirect perform_sum, N=%lld \n", N);
